Refactor EventRegisterCard into three new components
